requirements.txt
requests
lxml
graia-application-mirai